+++ packages/select/src/select.tsx
@@ -755,6 +755,11 @@ export default defineComponent({
       virtualRenderRef.value?.scrollTo(0, 1);
     };
 
+    // 处理扩展区域点击事件
+    const handleExtensionClick = (e: MouseEvent) => {
+      e.stopPropagation();
+    }
+
     return {
       t,
       selected,
@@ -820,6 +825,7 @@ export default defineComponent({
       preloadItemCount,
       virtualRenderRef,
       setSelected,
+      handleExtensionClick,
     };
   },
   render() {
@@ -1061,12 +1067,10 @@ export default defineComponent({
             <span>{this.curContentText}</span>
           </div>
         )}
-        <div
-          class={this.resolveClassName('select-content')}
-          v-show={this.isShowSelectContent}
-        >
+        <div class={this.resolveClassName('select-content')}>
           <div
             ref='scrollContainerRef'
+            v-show={this.isShowSelectContent}
             style={{ maxHeight: `${this.scrollHeight}px`, minHeight: `${this.minHeight}px` }}
             class={this.isEnableVirtualRender ? '' : this.resolveClassName('select-dropdown')}
             onScroll={this.handleScroll}
@@ -1090,7 +1094,7 @@ export default defineComponent({
             </ul>
           </div>
           {this.$slots?.extension && (
-            <div class={this.resolveClassName('select-extension')}>{this.$slots?.extension()}</div>
+            <div class={this.resolveClassName('select-extension')} onClick={this.handleExtensionClick}>{this.$slots?.extension()}</div>
           )}
         </div>
       </div>